function-source/requirements.txt
functions-framework==3.9.2
google-auth-httplib2==0.2.0
google-cloud-storage==2.16.0
requests==2.31.0
google-api-python-client==2.183.0
google-auth==2.40.3
google-cloud-run==0.11.0